# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "e-Disaster API",
        "description": "Disaster Management System API for Mobile Applications",
        "contact": {
            "email": "admin@edisaster.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/disasters/{id}/reports": {
            "get": {
                "tags": ["Reports"],
                "summary": "Get disaster reports",
                "description": "Get paginated list of reports for a specific disaster (assigned users only)",
                "operationId": "2763be38f854d317c94e1da4fb2261c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Disaster ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0199cfbc-eab1-7262-936e-72f9a6c5f659"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for title or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "damage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - not assigned to disaster",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied. You are not assigned to this disaster."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Disaster not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Disaster not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Reports"],
                "summary": "Create disaster report",
                "description": "Create a new report for a specific disaster (assigned users only)",
                "operationId": "d5a72fe7a1ee8880d9d3d3f8267acb78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Disaster ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0199cfbc-eab1-7262-936e-72f9a6c5f659"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["title", "description"],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Damage Assessment Report"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Detailed report on building damage and casualties"
                                    },
                                    "is_final_stage": {
                                        "description": "Mark as final stage (completes disaster)",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Disaster report created successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - not assigned to disaster",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied. You are not assigned to this disaster."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}